+++ frontend/srcs/src/app/modules/challenge/components/challenging/challenging.component.ts
@@ -3,6 +3,7 @@ import { User } from 'src/app/models';
 import { UsersCache } from 'src/app/cache';
 import { ApiService } from 'src/app/services';
 import { UriConstants } from 'src/app/utils';
+import { ChallengeService } from 'src/app/services/challenge.service';
 
 @Component({
   selector: 'app-challenging',
@@ -11,17 +12,20 @@ import { UriConstants } from 'src/app/utils';
 })
 export class ChallengingComponent implements OnInit {
   @Input() challengingUserId!: string;
+
   user: User = {
     userId: '',
     username: '',
   }
+  theme = localStorage.getItem('theme') || 'dark';
 
   private cachedUsers = inject(UsersCache);
   private apiService = inject(ApiService);
-
+  private challengeService = inject(ChallengeService);
 
   ngOnInit(): void {
     this.user = this.cachedUsers.getUser(this.challengingUserId);
+    this.challengeService.getThemeSub().subscribe(theme => { this.theme = theme });
   }
 
   cancelChallenge(){